Applied patch from Pu Long: ====================================================================== In ltp-20041007,gf15 and gf18 failed on both 32-bits and 64-bits platform(Tiger and normal i386 PC). Here is the reason: 		 Growfile used lseek and fstat to operate file. When file grows beyond 4G, lseek and fstat would fail on 32-bits machine. 		 I suggest replacing lseek with llseek and replacing fstat with another method which can handle 4G files correctly 		 The reason of failure on 64-bits machine is that growfile uses 'int' to hold offset/size of a file. The attached patch would fix that. 		 The failure of gf15 and gf18 is not stable and related to speed of disk I/O because they  grow/shrink files repeatedly for some certain time interval. 		 So within the same time interval, chances that a file grows beyond 4G on a machine with slower disk I/O are slimmer.
Pu Long

+++ testcases/kernel/fs/doio/growfiles.c
@@ -1838,14 +1838,14 @@ char *file;
 int grow_incr;
 char *buf;
 {
-   int noffset;
+   off_t noffset;
    int ret;
    int cur_offset;
    char *errmsg;
-   int fsize;		/* current size of file */
+   off_t fsize;		/* current size of file */
    int size_grew;	/* size the file grew */
    struct stat stbuf;
-   int tmp = 0;
+   off_t tmp = 0;
 
         /*
          * Do a stat on the open file.
